Move JexlValidator into jexl module
This way validators.py can be removed so there won't be any confusion
with new validation extension point
