+++ caluma/core/jexl.py
@@ -1,6 +1,17 @@
 import pyjexl
 from pyjexl.analysis import ValidatingAnalyzer
 from pyjexl.exceptions import ParseError
+from rest_framework import exceptions
+
+
+class JexlValidator(object):
+    def __init__(self, jexl):
+        self.jexl = jexl
+
+    def __call__(self, value):
+        errors = list(self.jexl.validate(value))
+        if errors:
+            raise exceptions.ValidationError(errors)
 
 
 class JEXL(pyjexl.JEXL):
